+++ src/itinerary/Itinerary.unit.tsx
@@ -79,6 +79,23 @@ describe('Itinerary component', () => {
     expect(itinerary.find('.kirk-itinerary-addon--to').exists()).toBe(false)
   })
 
+  describe('Should handle highlightRoad', () => {
+    it('Should add kirk-itinerary--highlightRoad class by default', () => {
+      const itinerary = shallow(<Itinerary places={places} />)
+      expect(itinerary.find('.kirk-itinerary--highlightRoad').exists()).toBe(true)
+    })
+
+    it('Should add kirk-itinerary--highlightRoad class if highighlightRoad is true', () => {
+      const itinerary = shallow(<Itinerary places={places} highighlightRoad />)
+      expect(itinerary.find('.kirk-itinerary--highlightRoad').exists()).toBe(true)
+    })
+
+    it('Should not add kirk-itinerary--highlightRoad class highighlightRoad is false', () => {
+      const itinerary = shallow(<Itinerary places={places} highlightRoad={false} />)
+      expect(itinerary.find('.kirk-itinerary--highlightRoad').exists()).toBe(false)
+    })
+  })
+
   it('Should display stopover with a11y label', () => {
     const itinerary = shallow(<Itinerary places={places} />)
     expect(itinerary.find('.kirk-itinerary-location')).toHaveLength(places.length)